+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @dyrected/nuxt
 
+## 2.5.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @dyrected/admin@2.5.47
+  - @dyrected/vue@2.5.47
+  - @dyrected/core@2.5.47
+  - @dyrected/sdk@2.5.47
+
 ## 2.5.46
 
 ### Patch Changes