+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dyrected/react
 
+## 2.5.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @dyrected/admin@2.5.47
+  - @dyrected/sdk@2.5.47
+
 ## 2.5.46
 
 ### Patch Changes