+++ packages/storage-local/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dyrected/storage-local
 
+## 2.5.47
+
+### Patch Changes
+
+- @dyrected/core@2.5.47
+
 ## 2.5.46
 
 ### Patch Changes